Parser: fix SpecialForm subclasses failing to be defined
The `id` kwarg passed to `__init_subclass__` does not get passed when
`slots=True` for the dataclass. Adding slots required the class to be
recreated, but none of the custom kwargs get passed when this happens.
In fact, when it does this I don't think it has any way to even know
what those custom kwargs were.